+++ src/main/java/janggi/camp/Camp.java
@@ -1,4 +1,4 @@
-package janggi;
+package janggi.camp;
 
 public enum Camp {
