+++ src/main/java/janggi/piece/Chariot.java
@@ -1,9 +1,8 @@
 package janggi.piece;
 
-import janggi.Camp;
-import janggi.PieceSymbol;
-import janggi.Point;
+import janggi.board.Point;
 import janggi.board.Board;
+import janggi.camp.Camp;
 import java.util.HashSet;
 import java.util.Set;
