+++ src/test/java/janggi/PointTest.java
@@ -3,6 +3,7 @@ package janggi;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import janggi.board.Point;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
