+++ src/test/java/janggi/piece/SoldierTest.java
@@ -2,9 +2,9 @@ package janggi.piece;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import janggi.Camp;
-import janggi.Point;
+import janggi.board.Point;
 import janggi.board.Board;
+import janggi.camp.Camp;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
